Fix linting errors "moved renamed check-auth controller to single index file and added linting script to package.json (#69)"
