+++ server/errors/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 import { Request, Response, NextFunction } from 'express';
 import CustomError from './custom-error';
