+++ server/index.ts
@@ -1,4 +1,5 @@
 import app from './app';
 
 const port = app.get('port');
+// eslint-disable-next-line no-console
 app.listen(port, () => console.log(`Running on http://localhost:${port}`));
